renamed to handshake-scanner, config
