+++ packages/observation-extractors/handshake-observation-extractor/lib/types.ts
@@ -1,0 +1,13 @@
+export interface HandshakeTxOutput {
+  value: bigint;
+  address: {
+    version: number;
+    hash: string;
+    string: string;
+  };
+  covenant: {
+    type: number;
+    action: string;
+    items: string[];
+  };
+}
