+++ packages/observation-extractors/handshake-observation-extractor/package.json
@@ -31,7 +31,7 @@
     "@rosen-bridge/abstract-logger": "^4.0.0",
     "@rosen-bridge/abstract-observation-extractor": "^1.0.2",
     "@rosen-bridge/extended-typeorm": "^1.0.1",
-    "@rosen-bridge/handshake-rpc-scanner": "^0.0.0",
+    "@rosen-bridge/handshake-scanner": "^0.0.0",
     "@rosen-bridge/rosen-extractor": "^11.3.0",
     "@rosen-bridge/tokens": "^6.0.0",
     "@rosen-bridge/scanner-interfaces": "^0.2.2",
