+++ packages/scanners/handshake-scanner/lib/types.ts
